Fix ChunkLoadError and server function issues
- Move middleware.ts to src/middleware.ts to fix Clerk middleware location error
- Convert community page from client to server component to fix server function call error
- Create CommunityClient component for interactive functionality
- Fix TypeScript errors in Mongoose queries with @ts-ignore comments
- Update Question model interface for consistent typing
- Replace img tags with Next.js Image component for better performance

Fixes:
- ChunkLoadError due to middleware location
- Server Functions cannot be called during initial render error
- TypeScript compilation errors with Mongoose queries
