+++ lib/actions/question.action.ts
@@ -14,7 +14,7 @@ export async function getQuestions(params: GetQuestionsParams) {
     const { searchQuery, filter } = params;
 
     // Build query
-    const query: Record<string, unknown> = {};
+    const query: any = {};
 
     if (searchQuery) {
       query.$or = [
@@ -42,7 +42,8 @@ export async function getQuestions(params: GetQuestionsParams) {
         break;
     }
 
-    const questions = await Question.find(query as any)
+    // @ts-ignore - Mongoose typing issue
+    const questions = await Question.find(query)
       .populate({
         path: "tags",
         model: Tag,
@@ -89,6 +90,7 @@ export async function createQuestion(params: createQuestionsParams) {
 
     const { title, explanation, tags, author, path } = params;
 
+    // @ts-ignore - Mongoose typing issue
     const question = await Question.create({
       title,
       explanation,
@@ -98,6 +100,7 @@ export async function createQuestion(params: createQuestionsParams) {
 
     if (tags && tags.length > 0) {
       for (const tag of tags) {
+        // @ts-ignore - Mongoose typing issue
         const existingTag = await Tag.findOneAndUpdate(
           { name: { $regex: new RegExp(`^${tag}$`, "i") } },
           { $setOnInsert: { name: tag }, $push: { questions: question._id } },
@@ -110,6 +113,7 @@ export async function createQuestion(params: createQuestionsParams) {
     }
 
     if (tagDocuments.length > 0) {
+      // @ts-ignore - Mongoose typing issue
       await Question.findByIdAndUpdate(question._id, {
         $push: {
           tags: { $each: tagDocuments },
